Read desc
Added new command LibsDisguises
Tells the command user who made the plugin.
And if the player who called the command is called by the attractive
name 'libraryaddict' then it also tells that curious user the version of
the plugin.
Worried about it being a exploit? Why? I know of all bugs/exploits and
could just try them all out if I didn't know the version. If there was a
exploit in the first place anyways.
The command is there to give credit to me, and so I can urge servers to
update.

Against this command?
I see you don't like plugin developers getting credit for their work
that you use for free.
Please move back to disguisecraft.